requirements.txt
Django==5.1.6
djangorestframework==3.15.2
channels[daphne]==4.1.0
channels-redis==4.2.0
psycopg==3.2.5
psycopg2==2.9.10
uwsgi==2.0.25.1
python-decouple==3.8
django-redisboard==8.4.0
asgiref==3.8.1
Pillow==11.1.0
sqlparse==0.5.3
django-braces==1.16.0
django-embed-video==1.4.10
pymemcache==4.0.0
django-debug-toolbar==5.0.1
redis==5.2.1
requests==2.32.3